src/main/java/com/doityourself/workshop/features/project/media/exceptionHandler/GetProjectLocalResourceGlobalExceptionHandler.java
package com.doityourself.workshop.features.project.media.exceptionHandler;

import com.doityourself.workshop.common.command.CommandDTO;
import com.doityourself.workshop.common.command.ICommandExceptionHandler;
import com.doityourself.workshop.features.project.media.exception.GetProjectLocalResourceFailedException;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import java.util.Collections;

/**
 * Get Project Local Resource Global Exception Handler
 */
@Slf4j
@Component
public class GetProjectLocalResourceGlobalExceptionHandler implements ICommandExceptionHandler<CommandDTO> {
  @Value("${project.localresource.validation.failed}")
  String projectLocalResourceFailedErrorMessage;

  /**
   * Method to handle runtime exception
   *
   * @param dto {@link CommandDTO}
   * @param runtimeException {@link RuntimeException}
   */
  @Override
  public void handleRuntimeException(CommandDTO dto, RuntimeException runtimeException) {
    handle();
  }

  /**
   * Method to handle exception
   *
   * @param dto {@link CommandDTO}
   * @param exception {@link Exception}
   */
  @Override
  public void handleException(CommandDTO dto, Exception exception) {
    handle();
  }

  /**
   * Common exception handler
   */
  private void handle() {
    GetProjectLocalResourceFailedException exception = new GetProjectLocalResourceFailedException();
    exception.setMessages(Collections.singletonList(projectLocalResourceFailedErrorMessage));
    throw exception;
  }
}
